feat(ddd): add user controllers test
